Add MD5 data type tests

+++ src/data-types/md5global.h
@@ -41,6 +41,7 @@
 #define MD5GLOBAL_H
 
 #include "md5namespace.h"
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -63,7 +64,7 @@ typedef const unsigned char *CONST_POINTER;
 typedef unsigned short int UINT2;
 
 /* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+typedef uint32_t UINT4;
 
 /* PROTO_LIST is defined depending on how PROTOTYPES is defined above.
 If using PROTOTYPES, then PROTO_LIST returns the list, otherwise it
